Move Booklist to containers
