+++ src/containers/BookList.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import Book from './Book';
+import Book from '../components/Book';
 
 const BookList = ({ books }) => (
   <div className="container">
